Vocal Barks / Voice Packs (#8162)
## About The Pull Request
Port of the vocal barks feature from a few other codebases, the core is
from https://github.com/Citadel-Station-13/Citadel-Station-13/pull/15677
and the rest has been entirely rewritten to *hopefully* work in the
monkestation environment without being obnoxious.

Each player can now select a voice pack. When a player talks they will
make a few 'barks' using it. The number of barks depends on the length
of their message.

There are a few rough edges but I'd like to have this playtested this
before refining it further.


https://github.com/user-attachments/assets/25d14b6a-8862-45c1-8065-5d5d733f5d82

Simple UI for previewing bark sounds utilizing some work from
https://github.com/Fluffy-Frontier/FluffySTG/pull/106

<img height="150" alt="image"
src="https://github.com/user-attachments/assets/3ae9c643-ba08-41d6-a875-00dff45181cc"
/>
<img height="270" alt="image"
src="https://github.com/user-attachments/assets/0c732001-201a-42c4-a70a-92a28b32dc0b"
/>

- Character Preferences
    - Voice Pack: What sound will the player make
    - Duration: How quickly does the player talk
- Pitch and Pitch Range: What is the pitch and variance of the player's
voice
- Game Preferences
- Single Bark: `When enabled, you will only hear a single sound when
players talk.` like the current system this can use a different sound
for exclaim, ask and talk
- Disable Pitch Modification: `When enabled, voice sounds will not play
with any character pitch modifications.`
- Only Goonstation: `When enabled, characters will only make goonstation
voice sounds.` if a character uses a non goonstation voice sound, one of
the default speak sounds will be used



- Voice Packs
    - TOML Config file
- Makes it very easy to add new voicepacks / improve on the existing
ones since the file can be reloaded at runtime
        - Volume control and limits for pitch / speed / whatever
- More goonstation talk sounds have been brought over, plus *some*
sounds from
https://github.com/Citadel-Station-13/Citadel-Station-13/pull/15684 +
https://github.com/Citadel-Station-13/Citadel-Station-13/pull/15677 +
https://github.com/Fluffy-Frontier/FluffySTG/pull/106


- Admin commands: Reload voice pack file, Toggle voices, Normalise
voicepack smite (sets the player to one of the default speak sounds)

- Integration
    - Changelings mimic the voice of whoever they're disguised as
- Players speaking in sign language will make a single snap bark unless
they are whispering
- Misc
- Split the settings button in the escape menu into game settings and
character settings

## Why It's Good For The Game
As with most of my PRs this makes character customisation more
effective.

This is part of a code bounty which several players wanted.

## Changelog
:cl:
add: Each player can now select a voice pack. When a player talks they
will make a few 'barks' using it. The number of barks depends on the
length of their message.
add: Players speaking in sign language will make a single snap bark
unless they are whispering
qol: Split the settings button in the escape menu into game settings and
character settings
/:cl:

---------

+++ code/modules/escape_menu/home_page.dm
@@ -15,9 +15,20 @@
 			null,
 			/* hud_owner = */ null,
 			src,
-			"Settings",
+			"Game Settings",
 			/* offset = */ 1,
-			CALLBACK(src, PROC_REF(home_open_settings)),
+			CALLBACK(src, PROC_REF(home_open_game_settings)),
+		)
+	)
+
+	page_holder.give_screen_object(
+		new /atom/movable/screen/escape_menu/home_button(
+			null,
+			/* hud_owner = */ null,
+			src,
+			"Edit Character",
+			/* offset = */ 2,
+			CALLBACK(src, PROC_REF(home_open_character_settings)),
 		)
 	)
 
@@ -27,7 +38,7 @@
 			/* hud_owner = */ null,
 			src,
 			"Redeem Code",
-			/* offset = */ 2,
+			/* offset = */ 3,
 			CALLBACK(src, PROC_REF(start_redeem)),
 		)
 	)
@@ -37,7 +48,7 @@
 			/* hud_owner = */ null,
 			src,
 			"Open Lootbox",
-			/* offset = */ 3,
+			/* offset = */ 4,
 			CALLBACK(src, PROC_REF(try_open_lootbox)),
 		)
 	)
@@ -48,7 +59,7 @@
 			/* hud_owner = */ null,
 			src,
 			"Open Map",
-			/* offset = */ 4,
+			/* offset = */ 5,
 			CALLBACK(src, PROC_REF(open_map)),
 		)
 	)
@@ -58,7 +69,7 @@
 			/* hud_owner = */ null,
 			src,
 			"Admin Help",
-			/* offset = */ 5,
+			/* offset = */ 6,
 		)
 	)
 
@@ -68,7 +79,7 @@
 			/* hud_owner = */ null,
 			src,
 			"Leave Body",
-			/* offset = */ 6,
+			/* offset = */ 7,
 			CALLBACK(src, PROC_REF(open_leave_body)),
 		)
 	)
@@ -87,8 +98,24 @@
 	if(url && client)
 		client << link(url)
 
-/datum/escape_menu/proc/home_open_settings()
-	client?.prefs.ui_interact(client?.mob)
+/datum/escape_menu/proc/home_open_game_settings()
+	var/datum/preferences/preferences = usr?.client?.prefs
+	if (!preferences)
+		return
+
+	preferences.current_window = PREFERENCE_TAB_GAME_PREFERENCES
+	preferences.update_static_data(usr)
+	preferences.ui_interact(usr)
+	qdel(src)
+
+/datum/escape_menu/proc/home_open_character_settings()
+	var/datum/preferences/preferences = usr?.client?.prefs
+	if (!preferences)
+		return
+
+	preferences.current_window = PREFERENCE_TAB_CHARACTER_PREFERENCES
+	preferences.update_static_data(usr)
+	preferences.ui_interact(usr)
 	qdel(src)
 
 /atom/movable/screen/escape_menu/home_button
